Update result models/interfaces
