+++ src/Result/AbstractResult.php
@@ -6,19 +6,11 @@ abstract class AbstractResult implements ResultInterface
 {
     private $isValid;
     private $model;
-    private $type;
-    private $code;
 
-    public function __construct(
-        bool $isValid,
-        object $model,
-        int $type = TypeInterface::NOT_APPLICABLE,
-        int $code = 0
-    ) {
+    public function __construct(bool $isValid, object $model)
+    {
         $this->isValid = $isValid;
         $this->model = $model;
-        $this->type = $type;
-        $this->code = $code;
     }
 
     public function getIsValid(): bool
@@ -29,15 +21,5 @@ abstract class AbstractResult implements ResultInterface
     public function getModel(): object
     {
         return $this->model;
-    }
-
-    public function getType(): int
-    {
-        return $this->type;
-    }
-
-    public function getCode(): int
-    {
-        return $this->code;
     }
 }
