+++ src/Result/InvalidResultInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace webignition\BasilModelValidator\Result;
+
+interface InvalidResultInterface extends ResultInterface
+{
+    public function getType(): int;
+    public function getCode(): int;
+}
