+++ src/Result/ResultInterface.php
@@ -6,6 +6,4 @@ interface ResultInterface
 {
     public function getIsValid(): bool;
     public function getModel(): object;
-    public function getType(): int;
-    public function getCode(): int;
 }
