+++ tests/Unit/Result/ValidResultTest.php
@@ -4,7 +4,6 @@
 namespace webignition\BasilModelValidator\Tests\Unit\Result;
 
 use webignition\BasilModel\Action\ActionInterface;
-use webignition\BasilModelValidator\Result\TypeInterface;
 use webignition\BasilModelValidator\Result\ValidResult;
 
 class ValidResultTest extends \PHPUnit\Framework\TestCase
@@ -17,7 +16,5 @@ class ValidResultTest extends \PHPUnit\Framework\TestCase
 
         $this->assertTrue($result->getIsValid());
         $this->assertSame($model, $result->getModel());
-        $this->assertEquals(TypeInterface::NOT_APPLICABLE, $result->getType());
-        $this->assertEquals(0, $result->getCode());
     }
 }
